+++ scripts/monster.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "monster.hpp"
 
-monster::monster(std::string name, int hp, int atk, int def, int add, int crit, int dodge): entity(name, hp, atk, def, add, crit, dodge)
+monster::monster(std::string name, float hp, float atk, float def, int add, int crit, int dodge): entity(name, hp, atk, def, add, crit, dodge)
 {
 	_name = name;
 	_hp = hp;
